backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
beautifulsoup4
lxml
faiss-cpu
numpy
openai
google-generativeai
langchain
langchain-community
tiktoken
aiofiles
python-multipart
redis
celery
motor
pymongo
prometheus-client

